Realizando acertos no Cadastro e LIstagem de Fornecedores

+++ server/estoque/src/main/java/br/com/gestaoLaboratorio/estoque/persistence/entity/Fornecedor.java
@@ -26,7 +26,7 @@ public class Fornecedor {
 
     private String nomeContato;
 
-    private int telefoneContato;
+    private Long telefoneContato;
 
     private String cnpj;
